reintroduced the credentials module - speeds things way up
