+++ postgres_query.py
@@ -9,49 +9,9 @@ from dotenv import load_dotenv
 from azure.identity import DefaultAzureCredential
 from azure.keyvault.secrets import SecretClient
 
-def sql_engine_string_generator(datahub_host, datahub_db, datahub_user, datahub_pwd): 
-
-    # set a try except clause to grab the online credentials keys and if not, grab them locally as environment variables
-    try:
-        # set the key vault path
-        KEY_VAULT_URL = "https://fsdh-proj-swapit-poc-kv.vault.azure.net/"
-        error_occur = False
-
-        # Retrieve the secrets containing DB connection details
-        credential = DefaultAzureCredential()
-        secret_client = SecretClient(vault_url=KEY_VAULT_URL, credential=credential)
-
-        # Retrieve the secrets containing DB connection details
-        DB_HOST = secret_client.get_secret(datahub_host).value
-        DB_NAME = secret_client.get_secret(datahub_db).value
-        DB_USER = secret_client.get_secret(datahub_user).value
-        DB_PASS = secret_client.get_secret(datahub_pwd).value
-        print ('Credentials loaded from FSDH')
-
-    except Exception as e:
-        # declare FSDH keys exception
-        error_occur = True
-        # print(f"An error occurred: {e}")
-
-        # load the .env file using the dotenv module remove this when running a powershell script to confirue system environment vars
-        parent_dir=os.path.dirname(os.getcwd())
-        load_dotenv(os.path.join(parent_dir, '.env')) # default is relative local directory 
-        DB_HOST = os.getenv(datahub_host)
-        DB_NAME = os.getenv(datahub_db)
-        DB_USER = os.getenv(datahub_user)
-        DB_PASS = os.getenv(datahub_pwd)
-        # print ('Credentials loaded locally')
+def fig_generator(start_date,end_date,sql_query,sql_engine_string):
 
     # set the sql engine string
-    sql_engine_string=('postgresql://{}:{}@{}/{}?sslmode=require').format(DB_USER,DB_PASS,DB_HOST,DB_NAME)
-    # print ('sql engine string: ',sql_engine_string)
-    return sql_engine_string
-
-
-def fig_generator(start_date,end_date,sql_query, database_name):
-
-    # set the sql engine string
-    sql_engine_string=sql_engine_string_generator('DATAHUB_PSQL_SERVER',database_name,'DATAHUB_PSQL_USER','DATAHUB_PSQL_PASSWORD')
     sql_engine=create_engine(sql_engine_string)
     conn = sql_engine.connect()
 
@@ -125,14 +85,3 @@ def fig_generator(start_date,end_date,sql_query, database_name):
     conn.close()
     return fig
 
-def first_entry(table,database_name):
-    # set the sql engine string
-    sql_engine_string=sql_engine_string_generator('DATAHUB_PSQL_SERVER',database_name,'DATAHUB_PSQL_USER','DATAHUB_PSQL_PASSWORD')
-    print ('sql credentials', sql_engine_string)
-    sql_engine=create_engine(sql_engine_string)
-    conn = sql_engine.connect()
-
-    first_entry_query=('SELECT datetime from {};').format(table)
-    output = conn.execute(text(first_entry_query))
-    conn.close()
-    return output.fetchone()[0]
